Fix error in exercise 8 chapter 6.
I think the value of std deviation (sigma) should be 1 instead of 0.1 in the formulas above.

The X_k's are Poisson with mean and variance 1 (sigma = 1). When applying the Central Limit Theorem, the new variance std deviation is (sigma / sqrt(N)) = (1 / 10).

In the equation above we divide twice by 10. Once when we used 0.1 instead of 1, and then a second time by multiplying by sqrt(N) in the numerator.